[#30] refactor :recycle: DetailViewController에서 View 부분 분리
